+++ app/src/main/java/au/com/shiftyjelly/pocketcasts/PocketCastsApplication.kt
@@ -49,7 +49,8 @@ import java.util.concurrent.Executors
 import javax.inject.Inject
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.collectLatest
+import kotlinx.coroutines.flow.launchIn
+import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.withContext
@@ -258,11 +259,12 @@ class PocketCastsApplication : Application(), Configuration.Provider {
         downloadManager.beginMonitoringWorkManager(applicationContext)
         userManager.beginMonitoringAccountManager(playbackManager)
 
-        applicationScope.launch {
-            settings.useDynamicColorsForWidget.flow.collectLatest {
-                widgetManager.updateWidgetFromSettings(playbackManager)
-            }
-        }
+        settings.useDynamicColorsForWidget.flow
+            .onEach { widgetManager.updateWidgetFromSettings(playbackManager) }
+            .launchIn(applicationScope)
+        settings.useRssArtwork.flow
+            .onEach { widgetManager.updateWidgetRssArtwork(playbackManager) }
+            .launchIn(applicationScope)
 
         Timber.i("Launched ${BuildConfig.APPLICATION_ID}")
     }